Started working on the actual pcap execution
Also fixed a bug in the config parser related to multiple interfaces/configs.
The module interface currently doesn't support any form of callbacks and the
pcap callback only prints the length of the packet for now. However the actual
module callback will probably be implemented in the next few commits.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>