Got rid of the interface struct and the range parameter
We're just saving the interface as a char* in the config struct now.
And I got rid of the range parameter as it wasn't used anyway. It is
probably a better idea to just have this as a module option if a module
decides it wants to use an ip range.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>