Added a preCapture function to the module interface
As this will get the event_base you can init other libevent stuff in here
like a timer for example. If you decided to implement this function (not
required) you are required to return a positive int from here, so yes you
can still prevent it from actually running in here. Due to the current
implementation however this is not recommended.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>