Went from the database pool model to a single connection model
Modules will have to handle the database connection themselves,
in most cases 1 will be plenty. If 1 is not enough simply make
more, you'll have to load balance them yourself though. I pretty
much went for this model instead of the old one because I had some
annoying race conditions in the database with the old model.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>