Changed the interval at which the database timer runs
100 milliseconds kind of meant the queue would fill up too quick
in some cases. It does not seem to do that with 100 nanoseconds.
Yes this causes a bit more load on your system, but on the other
side if you're using some rule on the database that uses now() those
are actually accurate. And you won't have the database queue overflowing
like crazy. Sure it can probably still peak up to 100+ but it'll (hopefully)
quickly execute all of them. Just don't do anything crazy to heavily delay
your queries.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>