Started on reworking the database
As it turns out smnl used in my case quite some memory. Turns out
this is because of me using the ipv4 module to keep track of bandwidth
usage. This fires so much queries at the database that it can't keep
up, resulting in the query queue of smnl just filling up. Therefor I'm
going to rework some things with the database. To begin libevent monitors
PQsocket() now, so no more need for that timer. Not on master yet as it
no longer handles database reconnection because of this. Regarding the
ipv4 module issue with a lot of data, I'll likely implement some sort of
hashmap which keeps track of stuff locally and just inserts it in the
database every so often, this will require some changes to the module API
however.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>