Keeping track of bandwidth per hosts locally now, inserting it every 10 seconds
We now have a simple binary search tree in which we store the
hosts and keep track of the bandwidth. We now loop through this
every 10 seconds and launch the actual queries, just after this
we of course clear it so the local 'cache' of it is reset. Currently
hardcoded to 10 seconds, will probably make this a config option in
a future commit.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>